refactor(provider): split transform by family
